Agrega funcionalidad de clientes, consulta api de stadística poblacional, cálculo de promedio y desviación estandar

+++ src/main/java/review/intercorp/client/model/KpiClienteResponse.java
@@ -13,5 +13,5 @@ import lombok.Setter;
 public class KpiClienteResponse {
 
     private Float average;
-    private Float standardDesviation;
+    private Double standardDesviation;
 }
